fixed compilation errors

+++ include/genn/genn/currentSourceModels.h
@@ -120,8 +120,8 @@ class PoissonExp : public Base
     SET_PARAM_NAMES({"weight", "tauSyn", "rate"});
     SET_VARS({{"current", "scalar"}});
     SET_DERIVED_PARAMS({
-        {"ExpDecay", [](const auto &pars, double dt){ return std::exp(-dt / pars.at("tauSyn")); }},
-        {"Init", [](const auto &pars, double dt){ return pars.at("weight") * (1.0 - std::exp(-dt / pars.at("tauSyn"))) * (pars.at("tauSyn") / dt); }},
-        {"ExpMinusLambda", [](const auto &pars, double dt){ return std::exp(-(pars.at("rate") / 1000.0) * dt); }}});
+        {"ExpDecay", [](const auto &pars, double dt){ return std::exp(-dt / pars.at("tauSyn").cast<double>()); }},
+        {"Init", [](const auto &pars, double dt){ return pars.at("weight").cast<double>() * (1.0 - std::exp(-dt / pars.at("tauSyn").cast<double>())) * (pars.at("tauSyn").cast<double>() / dt); }},
+        {"ExpMinusLambda", [](const auto &pars, double dt){ return std::exp(-(pars.at("rate").cast<double>() / 1000.0) * dt); }}});
 };
 } // GeNN::CurrentSourceModels
